agents/fleet_analyzer/a2a_requirements.txt
google-cloud-aiplatform[adk,agent_engines]==1.90.0
google-cloud-spanner==3.54.0
google-genai==1.14.0
google-adk==0.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
fastapi==0.115.12
urllib3==2.4.0
a2a_common-0.1.0-py3-none-any.whl

